Minor restructuring
